refactor: Separated media/index into different packages
Also added mutex for classifier cache so that it can classify
asynchronously soon enouhg
