+++ internal/media/storage/storage_json.go
@@ -1,14 +1,13 @@
-package media
+package storage
 
 import (
 	"context"
 	"crypto/sha256"
 	"encoding/json"
 	"fmt"
-	"net/http"
 	"os"
 	"path"
-	"strings"
+	"sync"
 
 	"github.com/baalimago/clai/pkg/text/models"
 	"github.com/baalimago/go_away_boilerplate/pkg/ancli"
@@ -18,7 +17,7 @@ import (
 
 type subtitleStreamFinder interface {
 	// fid the media info for some item
-	find(item model.Item) (MediaInfo, error)
+	find(item model.Item) (model.MediaInfo, error)
 }
 
 type subtitleStreamExtractor interface {
@@ -30,6 +29,7 @@ type subtitleStreamExtractor interface {
 
 type jsonStore struct {
 	storePath          string
+	cacheMu            *sync.RWMutex
 	cache              map[string]model.Item
 	subStreamFinder    subtitleStreamFinder
 	subStreamExtractor subtitleStreamExtractor
@@ -64,7 +64,7 @@ func WithStorePath(storePath string) JSONStoreOption {
 
 func NewJSONStore(opts ...JSONStoreOption) *jsonStore {
 	subUtils := &ffmpegSubsUtil{
-		mediaCache: map[string]MediaInfo{},
+		mediaCache: map[string]model.MediaInfo{},
 	}
 
 	cfgDir, err := os.UserConfigDir()
@@ -78,6 +78,8 @@ func NewJSONStore(opts ...JSONStoreOption) *jsonStore {
 		subStreamFinder:    subUtils,
 		subStreamExtractor: subUtils,
 		storePath:          storePath,
+		cache:              make(map[string]model.Item),
+		cacheMu:            &sync.RWMutex{},
 		classifier: agent.NewClassifier(models.Configurations{
 			Model:     "gpt-5",
 			ConfigDir: claiPath,
@@ -120,6 +122,8 @@ func (s *jsonStore) loadPersistedItems(storeDirPath string) error {
 			continue
 		}
 		f.Close()
+		s.cacheMu.Lock()
+		defer s.cacheMu.Unlock()
 		for _, item := range items {
 			s.cache[item.ID] = item
 		}
@@ -134,9 +138,6 @@ func (s *jsonStore) Setup(ctx context.Context) error {
 
 	if _, err := os.Stat(s.storePath); err != nil {
 		os.MkdirAll(s.storePath, 0o755)
-	}
-	if s.cache == nil {
-		s.cache = make(map[string]model.Item)
 	}
 	err := s.loadPersistedItems(s.storePath)
 	if err != nil {
@@ -185,6 +186,8 @@ func generateID(i model.Item) string {
 }
 
 func (s *jsonStore) store(i model.Item) error {
+	s.cacheMu.Lock()
+	defer s.cacheMu.Unlock()
 	s.cache[i.ID] = i
 	storePath := path.Join(s.storePath, i.ID)
 	f, err := os.OpenFile(storePath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0o644)
@@ -218,7 +221,9 @@ func (s *jsonStore) Store(ctx context.Context, i model.Item) error {
 	if i.ID == "" {
 		i.ID = generateID(i)
 	}
+	s.cacheMu.RLock()
 	existingItem, exists := s.cache[i.ID]
+	s.cacheMu.RUnlock()
 
 	if exists {
 		// Only keep path if the item when it exists
@@ -244,130 +249,4 @@ func (s *jsonStore) Store(ctx context.Context, i model.Item) error {
 		}
 	}
 	return s.store(i)
-}
-
-// ListHandlerFunc returns a list of all available items in the gallery
-func (s *jsonStore) ListHandlerFunc() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if s.cache == nil {
-			http.Error(w, "store not initialized", http.StatusInternalServerError)
-			return
-		}
-		items := make([]model.Item, 0, len(s.cache))
-		for _, v := range s.cache {
-			items = append(items, v)
-		}
-		w.Header().Set("Content-Type", "application/json")
-		if err := json.NewEncoder(w).Encode(items); err != nil {
-			http.Error(w, "failed to encode items", http.StatusInternalServerError)
-		}
-	}
-}
-
-// VideoHandlerFunc returns a handler to get a video by ID, if item is not a video
-// it will return 404
-func (s *jsonStore) VideoHandlerFunc() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("id")
-		if id == "" {
-			http.Error(w, "missing id", http.StatusBadRequest)
-			return
-		}
-		item, ok := s.cache[id]
-		if !ok {
-			http.NotFound(w, r)
-			return
-		}
-		pathToMedia := item.Path
-		mimeType := item.MIMEType
-
-		if !strings.Contains(mimeType, "video") {
-			http.Error(w, "media found, but its not a video", http.StatusNotFound)
-			return
-		}
-
-		w.Header().Set("Content-Type", mimeType)
-		file, err := os.Open(pathToMedia)
-		if err != nil {
-			http.Error(w, "media not found", http.StatusNotFound)
-			return
-		}
-		defer file.Close()
-
-		info, err := os.Stat(pathToMedia)
-		if err != nil {
-			http.Error(w, "media not found", http.StatusNotFound)
-			return
-		}
-
-		// We'll see how robus this is. This rule covers all of my devices!
-		if strings.HasSuffix(strings.ToLower(item.Name), ".mkv") && !strings.Contains(r.UserAgent(), "SmartTV") {
-			streamMkvToMp4(w, r, pathToMedia)
-			return
-		}
-
-		modTime := info.ModTime()
-		http.ServeContent(w, r, item.Name, modTime, file)
-	}
-}
-
-// SubsHandlerFunc by stripping out the substitle streams using ffmpeg from video media found at
-// PathValue id. If there are multiple subtitle streams found, select one at random
-func (s *jsonStore) SubsListHandlerFunc() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("vid")
-		if id == "" {
-			http.Error(w, "missing vid", http.StatusBadRequest)
-			return
-		}
-		item, ok := s.cache[id]
-		if !ok {
-			http.NotFound(w, r)
-			return
-		}
-		if !strings.Contains(item.MIMEType, "video") {
-			http.Error(w, "media found, but its not a video", http.StatusNotFound)
-			return
-		}
-
-		info, err := s.subStreamFinder.find(item)
-		if err != nil {
-			ancli.Errf("jsonStore failed to handle SubsList when subStripper.extract, err: %v", err)
-			http.Error(w, "failed to extract subtitles from media", http.StatusInternalServerError)
-			return
-		}
-		w.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(w).Encode(info)
-		ancli.Okf("media info served")
-	}
-}
-
-func (s *jsonStore) SubsHandlerFunc() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vid := r.PathValue("vid")
-		if vid == "" {
-			http.Error(w, "missing video id", http.StatusBadRequest)
-			return
-		}
-
-		sid := r.PathValue("sub_idx")
-		if sid == "" {
-			http.Error(w, "missing subtitle index", http.StatusBadRequest)
-			return
-		}
-
-		cacheFile, exists := s.cache[vid]
-		if !exists {
-			http.Error(w, fmt.Sprintf("cache miss for: '%v'", vid), http.StatusNotFound)
-			return
-		}
-		subs, err := s.subStreamExtractor.extract(cacheFile, sid)
-		if err != nil {
-			ancli.Errf("failed to extract subs: %v", err)
-			http.Error(w, "failed to extract subs", http.StatusInternalServerError)
-			return
-		}
-		w.Header().Set("Content-Type", "text/vtt; charset=utf-8")
-		http.ServeFile(w, r, subs)
-	}
 }
